pyproject.toml
[project]
name = "personal-knowledge-assistant"
version = "0.1.0"
description = ""
requires-python = ">=3.10"

dependencies = [
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "requests>=2.32.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

